fix(zorder): one verified pass, never a second raise of a window
Device QA (owner, 2026-08-02): an App Bar click on the leftmost
item restacked the row correctly and then slid the focus one slot
back. Two things in the previous commits could raise a window
twice inside one sequence — the retry pass, and running a
coalesced restore from the drain's completion — and a second raise
is exactly what the raise-echo ledger cannot absorb.
`zOrderRaiseEchoes` clears a window's stamp on its FIRST echo
(`KiwiCore+FocusEvents`), so the second echo arrives unstamped, is
read as a deliberate focus change, and focus lands on the
pile-mate. The last window a sequence raises is the one next to
the focus, which is why the row moved by exactly one slot.

Both come out. The drain now makes exactly one verifying pass and
raises each window at most once; a window that misses its landing
is left for the next restore rather than retried, which
measurement supports — no raise was ever lost, only late. Arming
goes back to running immediately, so a restore during a drain
queues behind it as it did before; coalescing needs the ledger
re-stamped per sequence before it is safe, and that is #684's
second half, still open.

Also from the guard-prover run: the fake WindowServer modelled
every raise as reaching the front of the stacking, which is the
assumption device measurement falsified — a quiet raise never gets
above the frontmost app's key window (0 of 7 windows over 600 ms).
It now models that ceiling, which is what makes the float floor
testable at all: the floor deliberately excludes the focused
window, because a floor containing the key window is a condition
no raise can satisfy and the drain would burn its whole budget
failing it on every focus change.
